Task 7.15 - реализовано заполнение шаблона
